create a new file task.go and move task part codes from cluster.go into tasks.go
